requirements.txt
numpy==1.24.3
openai-whisper
Wave==0.0.2
openai
customtkinter==5.1.3
PyAudioWPatch==0.2.12.5
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchaudio
funasr
python-dotenv
modelscope 
huggingface_hub